Aside hero list can now be fetched

+++ src/main/java/com/hodor/dota2partner/dto/AsideHeroRequestDTO.java
@@ -12,7 +12,7 @@ public class AsideHeroRequestDTO {
     private String name;
     private String image;
     private int matchPlayed;
-    private double winRate;
+    private float winRate;
     private int averageKill;
     private int averageDeath;
     private int averageAssist;
